cat_fact_project/requirements.txt
Celery
Django
django-celery-beat 
django-celery-results # currently requires Django 4.0 or older
python-decouple
redis
requests
djangorestframework